+++ scripts/disable_local_dev.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 
+# Please make sure to run this script in the root directory of the repository and not inside sub-folders
+
 echo ""
 
-echo "Disable local development for flutter_quill:"
+echo "Disable local development for flutter_quill..."
 rm pubspec_overrides.yaml
 
 echo ""
 
-echo "Enable local development for flutter_quill_extensions:"
+echo "Enable local development for flutter_quill_extensions..."
 rm flutter_quill_extensions/pubspec_overrides.yaml
 
 echo ""
 
-echo "Enable local development for flutter_quill_test:"
+echo "Enable local development for flutter_quill_test..."
 rm flutter_quill_test/pubspec_overrides.yaml
 
 echo ""