+++ lib/src/utils/platform.dart
@@ -4,62 +4,113 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart'
     show kIsWeb, TargetPlatform, defaultTargetPlatform;
 
-bool isWeb() {
-  return kIsWeb;
+/// If you want to override the [kIsWeb] use [overrideIsWeb]
+bool isWeb({bool? overrideIsWeb}) {
+  return overrideIsWeb ?? kIsWeb;
 }
 
-bool isMobile([TargetPlatform? targetPlatform]) {
-  if (isWeb()) return false;
-  targetPlatform ??= defaultTargetPlatform;
-  return {TargetPlatform.iOS, TargetPlatform.android}.contains(targetPlatform);
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isMobile({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb) && !supportWeb) return false;
+  platform ??= defaultTargetPlatform;
+  return {TargetPlatform.iOS, TargetPlatform.android}.contains(platform);
 }
 
-bool isDesktop([TargetPlatform? targetPlatform]) {
-  if (isWeb()) return false;
-  targetPlatform ??= defaultTargetPlatform;
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isDesktop({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb) && !supportWeb) return false;
+  platform ??= defaultTargetPlatform;
   return {TargetPlatform.macOS, TargetPlatform.linux, TargetPlatform.windows}
-      .contains(targetPlatform);
+      .contains(platform);
 }
 
-bool isKeyboardOS([TargetPlatform? targetPlatform]) {
-  targetPlatform ??= defaultTargetPlatform;
-  return isDesktop(targetPlatform) || targetPlatform == TargetPlatform.fuchsia;
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isKeyboardOS({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  platform ??= defaultTargetPlatform;
+  return isDesktop(
+          platform: platform,
+          supportWeb: supportWeb,
+          overrideIsWeb: overrideIsWeb) ||
+      platform == TargetPlatform.fuchsia;
 }
 
-bool isAppleOS([TargetPlatform? targetPlatform]) {
-  if (isWeb()) return false;
-  targetPlatform ??= defaultTargetPlatform;
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isAppleOS({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb) && !supportWeb) return false;
+  platform ??= defaultTargetPlatform;
   return {
     TargetPlatform.macOS,
     TargetPlatform.iOS,
-  }.contains(targetPlatform);
+  }.contains(platform);
 }
 
-bool isMacOS([TargetPlatform? targetPlatform]) {
-  if (isWeb()) return false;
-  targetPlatform ??= defaultTargetPlatform;
-  return TargetPlatform.macOS == targetPlatform;
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isMacOS({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb) && !supportWeb) return false;
+  platform ??= defaultTargetPlatform;
+  return TargetPlatform.macOS == platform;
 }
 
-bool isIOS([TargetPlatform? targetPlatform]) {
-  if (isWeb()) return false;
-  targetPlatform ??= defaultTargetPlatform;
-  return TargetPlatform.iOS == targetPlatform;
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isIOS({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb) && !supportWeb) return false;
+  platform ??= defaultTargetPlatform;
+  return TargetPlatform.iOS == platform;
 }
 
-bool isAndroid([TargetPlatform? targetPlatform]) {
-  if (isWeb()) return false;
-  targetPlatform ??= defaultTargetPlatform;
-  return TargetPlatform.android == targetPlatform;
+/// [supportWeb] is a parameter that ask you if we should care about web support
+/// if the value is true then we will return the result no matter if we are
+/// on web or using a native app to run the flutter app
+bool isAndroid({
+  required bool supportWeb,
+  TargetPlatform? platform,
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb) && !supportWeb) return false;
+  platform ??= defaultTargetPlatform;
+  return TargetPlatform.android == platform;
 }
 
-bool isFlutterTest() {
-  if (isWeb()) return false;
-  return Platform.environment.containsKey('FLUTTER_TEST');
-}
-
-Future<bool> isIOSSimulator() async {
-  if (!isAppleOS()) {
+Future<bool> isIOSSimulator({
+  bool? overrideIsWeb,
+}) async {
+  if (!isAppleOS(supportWeb: false, overrideIsWeb: overrideIsWeb)) {
     return false;
   }
 
@@ -72,4 +123,11 @@ Future<bool> isIOSSimulator() async {
     return !iosInfo.isPhysicalDevice;
   }
   return false;
+}
+
+bool isFlutterTest({
+  bool? overrideIsWeb,
+}) {
+  if (isWeb(overrideIsWeb: overrideIsWeb)) return false;
+  return Platform.environment.containsKey('FLUTTER_TEST');
 }