+++ lib/src/widgets/delegate.dart
@@ -331,7 +331,9 @@ class EditorTextSelectionGestureDetectorBuilder {
   @protected
   void onDragSelectionEnd(DragEndDetails details) {
     renderEditor!.handleDragEnd(details);
-    if (isDesktop() &&
+    // TODO: Should we care if the platform is desktop using native desktop app
+    // or the flutter app is running using web app??
+    if (isDesktop(supportWeb: true) &&
         delegate.selectionEnabled &&
         shouldShowSelectionToolbar) {
       // added to show selection copy/paste toolbar after drag to select